+++ index.js
@@ -4,9 +4,16 @@ const MODULE_REQUIRE = 1
     /* built-in */
     
     /* NPM */
+    , noda = require('noda')
     
     /* in-package */
+    , dualSwift = noda.inRequire('swift')
     ;
 
+function isConnection(conn) {
+    return conn instanceof dualSwift.Connection;
+}
 
-module.exports = '';
+module.exports = {
+    isConnection,
+};