+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 Notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning 2.0.0](http://semver.org/).
 
+##  [0.0.3] - Dec 9th, 2018
+
+*   Change the entrance (property "main" in [package.json](./package.json)) from `swift/index` to `index`.
+*   Method `isConnection()` added.
+*   Method `get()` added to class `ceph-dual/swift.Connection`.
+*   Setter `.container` added to class `ceph-dual/swift.Connection`.
+
 ##	[0.0.1] - 2018-9-4
 
 Released.